+++ src/main/kotlin/com/emo_hip_hop/mz2mo/account/domain/AccountId.kt
@@ -1,0 +1,5 @@
+package com.emo_hip_hop.mz2mo.account.domain
+
+data class AccountId(
+    val id: String
+)
