--- src/main/kotlin/com/emo_hip_hop/mz2mo/domain/account/domain/AccountId.kt
@@ -1,5 +1,0 @@
-package com.emo_hip_hop.mz2mo.domain.account.domain
-
-data class AccountId(
-    val id: String
-)
