--- src/main/kotlin/com/emo_hip_hop/mz2mo/domain/emoji/domain/EmojiId.kt
@@ -1,5 +1,0 @@
-package com.emo_hip_hop.mz2mo.domain.emoji.domain
-
-data class EmojiId(
-    val id: String
-)
