--- src/main/kotlin/com/emo_hip_hop/mz2mo/domain/music/domain/MusicArticleId.kt
@@ -1,5 +1,0 @@
-package com.emo_hip_hop.mz2mo.domain.music.domain
-
-data class MusicArticleId(
-    val id: String
-)
