+++ src/main/kotlin/com/emo_hip_hop/mz2mo/emoji/domain/Emoji.kt
@@ -1,4 +1,4 @@
-package com.emo_hip_hop.mz2mo.domain.emoji.domain
+package com.emo_hip_hop.mz2mo.emoji.domain
 
 data class Emoji(
     val id: EmojiId,
