+++ src/main/kotlin/com/emo_hip_hop/mz2mo/emoji/domain/EmojiId.kt
@@ -1,0 +1,5 @@
+package com.emo_hip_hop.mz2mo.emoji.domain
+
+data class EmojiId(
+    val id: String
+)
