+++ src/main/kotlin/com/emo_hip_hop/mz2mo/music/application/port/input/AddMusicArticleUseCase.kt
@@ -1,0 +1,7 @@
+package com.emo_hip_hop.mz2mo.music.application.port.input
+
+import com.emo_hip_hop.mz2mo.music.domain.MusicArticle
+
+interface AddMusicArticleUseCase {
+    operator fun invoke(command: AddMusicCommand): MusicArticle
+}
