+++ src/main/kotlin/com/emo_hip_hop/mz2mo/music/application/port/input/AddMusicCommand.kt
@@ -1,4 +1,4 @@
-package com.emo_hip_hop.mz2mo.domain.music.application.port.input
+package com.emo_hip_hop.mz2mo.music.application.port.input
 
 import com.emo_hip_hop.mz2mo.global.validate.SelfValidating
 import javax.validation.constraints.NotBlank
