+++ src/main/kotlin/com/emo_hip_hop/mz2mo/music/application/port/output/CreateMusicArticlePort.kt
@@ -1,0 +1,7 @@
+package com.emo_hip_hop.mz2mo.music.application.port.output
+
+import com.emo_hip_hop.mz2mo.music.domain.MusicArticle
+
+interface CreateMusicArticlePort {
+    fun create(musicArticle: MusicArticle): MusicArticle
+}
