+++ src/main/kotlin/com/emo_hip_hop/mz2mo/music/application/port/output/ExistsMusicArticlePort.kt
@@ -1,4 +1,4 @@
-package com.emo_hip_hop.mz2mo.domain.music.application.port.output
+package com.emo_hip_hop.mz2mo.music.application.port.output
 
 interface ExistsMusicArticlePort {
     fun byId(youtubeTrackId: String): Boolean
