+++ src/main/kotlin/com/emo_hip_hop/mz2mo/music/application/service/AddMusicArticleService.kt
@@ -1,11 +1,11 @@
-package com.emo_hip_hop.mz2mo.domain.music.application.service
+package com.emo_hip_hop.mz2mo.music.application.service
 
-import com.emo_hip_hop.mz2mo.domain.music.application.port.input.AddMusicArticleUseCase
-import com.emo_hip_hop.mz2mo.domain.music.application.port.input.AddMusicCommand
-import com.emo_hip_hop.mz2mo.domain.music.application.port.output.CreateMusicArticlePort
-import com.emo_hip_hop.mz2mo.domain.music.application.port.output.ExistsMusicArticlePort
-import com.emo_hip_hop.mz2mo.domain.music.domain.MusicArticle
-import com.emo_hip_hop.mz2mo.domain.music.domain.MusicVotes
+import com.emo_hip_hop.mz2mo.music.application.port.input.AddMusicArticleUseCase
+import com.emo_hip_hop.mz2mo.music.application.port.input.AddMusicCommand
+import com.emo_hip_hop.mz2mo.music.application.port.output.CreateMusicArticlePort
+import com.emo_hip_hop.mz2mo.music.application.port.output.ExistsMusicArticlePort
+import com.emo_hip_hop.mz2mo.music.domain.MusicArticle
+import com.emo_hip_hop.mz2mo.music.domain.MusicVotes
 import com.emo_hip_hop.mz2mo.global.UseCase
 
 @UseCase
