+++ src/main/kotlin/com/emo_hip_hop/mz2mo/music/application/service/MusicArticleAlreadyExistsException.kt
@@ -1,4 +1,4 @@
-package com.emo_hip_hop.mz2mo.domain.music.application.service
+package com.emo_hip_hop.mz2mo.music.application.service
 
 class MusicArticleAlreadyExistsException(val youtubeTrackId: String) : RuntimeException() {
