+++ src/main/kotlin/com/emo_hip_hop/mz2mo/music/domain/MusicArticle.kt
@@ -1,4 +1,4 @@
-package com.emo_hip_hop.mz2mo.domain.music.domain
+package com.emo_hip_hop.mz2mo.music.domain
 
 import java.util.*
