+++ src/main/kotlin/com/emo_hip_hop/mz2mo/music/domain/MusicArticleId.kt
@@ -1,0 +1,5 @@
+package com.emo_hip_hop.mz2mo.music.domain
+
+data class MusicArticleId(
+    val id: String
+)
